Struct google_cloudtrace1::TraceSpan

pub struct TraceSpan {
    pub kind: Option<String>,
    pub parent_span_id: Option<String>,
    pub name: Option<String>,
    pub start_time: Option<String>,
    pub span_id: Option<String>,
    pub labels: Option<HashMap<String, String>>,
    pub end_time: Option<String>,
}

A span represents a single timed event within a trace. Spans can be nested and form a trace tree. Often, a trace contains a root span that describes the end-to-end latency of an operation and, optionally, one or more subspans for its suboperations. Spans do not need to be contiguous. There may be gaps between spans in a trace.

This type is not used in any activity, and only used as part of another schema.

Fields

kind

Distinguishes between spans generated in a particular context. For example, two spans with the same name may be distinguished using RPC_CLIENT and RPC_SERVER to identify queueing latency associated with the span.

parent_span_id

ID of the parent span, if any. Optional.

name

Name of the trace. The trace name is sanitized and displayed in the Cloud Trace tool in the Google Developers Console. The name may be a method name or some other per-call site name. For the same executable and the same call point, a best practice is to use a consistent name, which makes it easier to correlate cross-trace spans.

start_time

Start time of the span in nanoseconds from the UNIX epoch.

span_id

Identifier for the span. This identifier must be unique within a trace.

labels

Collection of labels associated with the span.

end_time

End time of the span in nanoseconds from the UNIX epoch.